+++ docs/versionhistory.rst
@@ -3,6 +3,14 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
+**UNRELEASED**
+
+- Added support for ``typing.Protocol`` subclasses
+- Added support for ``typing.AbstractSet``
+- Fixed the handling of ``total=False`` in ``TypedDict``
+- Fixed no error reported on unknown keys with ``TypedDict``
+- Removed support of default values in ``TypedDict``, as they are not supported in the spec
+
 **2.6.1** (2019-11-17)
 
 - Added the possibility to emit ``TypeWarning`` instead of raising ``TypeError`` on type violations